Components details

+++ client/src/components/Cart/PostSale.jsx
@@ -1,50 +1,57 @@
 import axios from 'axios'
 import { useState, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import { useSearchParams } from 'react-router-dom'
+import { useNavigate, useSearchParams } from 'react-router-dom'
 import { loadCart } from '../../Redux/reducer/cartSlice';
 import './PostSale.css';
 import Carousel from '../Home/Carousel/Carousel';
 import LoaderBars from '../common/LoaderBars';
 import DeliveryProgress from '../common/DeliveryProgress';
 import { ReactComponent as Spinner } from '../../assets/svg/spinner.svg';
+
 const PostSale = () => {
     const [order, setOrder] = useState(false);
     const [loading, setLoading] = useState(true);
     const [images, setImages] = useState(false);
     const [background, setBackground] = useState('')
     const dispatch = useDispatch();
+    const navigate = useNavigate();
     
     const [params] = useSearchParams(),
     id = params.get('external_reference');
     let status = params.get('status') || 'cancelled';
     
     useEffect(() => {
-        (async () => {
-            if (status === 'null' || status === 'cancelled') {
-                setBackground('postsale-pending animation-start');
-                setTimeout(() => {
-                    setBackground('postsale-pending animation-loop');                    
-                }, 6000);
-                // return navigate('/')
-            };
-            
+        (async () => {            
             const { data } = await axios(`/order/${id}`);
+            if (!data.products) {
+                console.error('no order');
+                navigate('/')
+            }
             console.log(data);
             setOrder(data);
+            setLoading(false);
 
-            if (!order.payment_date && status !== 'approved') deliveryWaiter(id);
+
+            if (data && !data?.payment_date && status === 'approved') deliveryWaiter(id);
 
             let aux = [];
             data.products.forEach(e => {
                 aux.push({img: e.img, url: ''});
             });
             setImages(aux);
-
-            setLoading(false)
             
-            if (status === 'approved') {
+            if (data && status === 'approved') {
+                console.log('llega aca approved #1');
+                //? background effect
+                setBackground('postsale-approved animation-start');
+                setTimeout(() => {
+                    setBackground('postsale-approved animation-loop');                    
+                }, 6000);
+                console.log('llega aca approved #2');
+
                 if (data.order_type === 'cart') {
+                console.log('llega aca approved #3');
                     //? vaciar carrito
                     await axios.delete(`/cart/empty`);
                     
@@ -54,35 +61,47 @@ const PostSale = () => {
                     //? Quitar last_order en el carrito de la db
                     await axios.put(`/cart/order/`);
                 } else {
+                console.log('llega aca approved #3b');
                     //? vaciar el buynow
                     axios.post(`/cart/`, {product_id: ''});
                 }
+                console.log('llega aca approved #4');
                 //? Cambiar el estado a 'processing'
                 await axios.put(`/order/${id}`,{
                     status: 'processing'
                 });
+                
+                console.log('llega aca approved #5');
                 //? en el back (choNotif):
-                // cambia orden a pagada    
+                // cambia orden a pagada
                 // resta unidades de cada stock
-
-                //? background effect
-                setBackground('postsale-approved animation-start');
-                setTimeout(() => {
-                    setBackground('postsale-approved animation-loop');                    
-                }, 6000);
             };
-            if (status !== 'approved' && data.status !== 'approved') {
+
+            if (data && status !== 'approved' && data.status !== 'approved') {
+                console.log('llega aca not approved');
+
                 setBackground('postsale-pending animation-start');
                 setTimeout(() => {
                     setBackground('postsale-pending animation-loop');                    
                 }, 6000);
-                //? cambiar estado de la orden si el status no es aprobado en ninguno de los casos
-                await axios.put(`/order/${id}`,{
-                    status
-                });
+                //? si llega como null o cancelled cambiar estado a "pending"
+                if (status === 'null' || status === 'cancelled') {
+                    console.log('llega aca cambiar estado');
+
+                    await axios.put(`/order/${id}`,{
+                        status: 'pending'
+                    });
+                } else {
+                    console.log('llega aca cambiar estado');
+                    //? cambiar estado de la orden si el status no es aprobado en ninguno de los casos
+                    await axios.put(`/order/${id}`,{
+                        status
+                    });
+                }
             }
+
         })()
-      // eslint-disable-next-line
+      //eslint-disable-next-line
     }, [])
 
     const deliveryWaiter = async (id) => {
@@ -90,6 +109,7 @@ const PostSale = () => {
             console.log('pidiendo orden nueva');
             const { data } = await axios(`/order/${id}`);            
             if (data.payment_date) {                
+                console.log('orden nueva recibida');
                 setOrder(data);
             } else {
                 throw new Error('')
@@ -115,14 +135,14 @@ const PostSale = () => {
     
     return (
         <div className='postsale-container'>
-            { (loading && !order )
+            {(loading && !order)
                 ? <LoaderBars />
                 : <div className='postsale-inner'>
                     <div className={`postsale-header ${background}`}>
                         <div className="postsale-img-container">
-                            <Carousel images={images} interval={2500} pausable ={false} width="15vw" height='15vw'/>
+                            <Carousel images={images} interval={2500} pausable ={false} width="25vh" height="25vh"/>
                             <div className='card-image-back-style'></div>
-                        </div>                        
+                        </div>
                     </div>
                     <div className="postsale-details-container">
                         {status === 'approved' && <div>
@@ -130,22 +150,21 @@ const PostSale = () => {
                             <h3>{messageQuantity()}</h3>
                             <h3>Ahora estamos preparando el envío.</h3>
 
-                            {order.delivery_date 
+                            {order?.delivery_date 
                                 ? <DeliveryProgress order={order}/>
                                 : <Spinner />
                             }
                         </div>}
-                        {/* <p>Resumen: {order?.description}</p> */}
                         {status !== 'approved' && <div>
                             <h1>HUBO ALGÚN ERROR...</h1>
                             <h3>Pero podés retomar el pago!</h3>
                             <button className='g-white-button'>
-                                <a href={order.payment_link}>Voy a intentarlo de nuevo</a>
+                                <a href={order?.payment_link}>Voy a intentarlo de nuevo</a>
                             </button>
                             <p className='postsale-text-margin'>También podés encontrar el link de pago en tu perfil para realizarlo más tarde.</p>
                         </div>}
                         <p>{`Estado del pago: ${status}`}</p>
-                        <p>Medio de pago: {order.payment_source}</p>
+                        <p>Medio de pago: {order?.payment_source}</p>
                         <p>Id de orden: <i>{order?.id}</i></p>
                     </div>
                 </div>}
@@ -154,3 +173,5 @@ const PostSale = () => {
 }
 
 export default PostSale
+
+
